Refactored the code to reach a threshold under .2 seconds
First test took well over the time it should have. The function that had to be refactored was the find anagrams function. It started off as object comprehensions and list comprehensions and a for loop. We scrapped the object comprehension for a for loop. We then we able to refactor again to git rid of the loops for a more concise way of doing it. We ended with a threshold of below a .3.